move parser mentioned core functions to duckdb main
