+++ extension/core_functions/function_list.cpp
@@ -67,7 +67,6 @@ static const StaticFunctionDefinition core_functions[] = {
 	DUCKDB_SCALAR_FUNCTION_SET(AgeFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(AggregateFun),
 	DUCKDB_SCALAR_FUNCTION(AliasFun),
-	DUCKDB_SCALAR_FUNCTION_ALIAS(ApplyFun),
 	DUCKDB_AGGREGATE_FUNCTION(ApproxCountDistinctFun),
 	DUCKDB_AGGREGATE_FUNCTION_SET(ApproxQuantileFun),
 	DUCKDB_AGGREGATE_FUNCTION(ApproxTopKFun),
@@ -82,14 +81,12 @@ static const StaticFunctionDefinition core_functions[] = {
 	DUCKDB_AGGREGATE_FUNCTION_ALIAS(ArrayAggFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayAggrFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayAggregateFun),
-	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayApplyFun),
 	DUCKDB_SCALAR_FUNCTION_SET(ArrayCosineDistanceFun),
 	DUCKDB_SCALAR_FUNCTION_SET(ArrayCosineSimilarityFun),
 	DUCKDB_SCALAR_FUNCTION_SET(ArrayCrossProductFun),
 	DUCKDB_SCALAR_FUNCTION_SET(ArrayDistanceFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayDistinctFun),
 	DUCKDB_SCALAR_FUNCTION_SET_ALIAS(ArrayDotProductFun),
-	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayFilterFun),
 	DUCKDB_SCALAR_FUNCTION_SET_ALIAS(ArrayGradeUpFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayHasAllFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayHasAnyFun),
@@ -100,7 +97,6 @@ static const StaticFunctionDefinition core_functions[] = {
 	DUCKDB_SCALAR_FUNCTION_SET_ALIAS(ArrayReverseSortFun),
 	DUCKDB_SCALAR_FUNCTION_SET_ALIAS(ArraySliceFun),
 	DUCKDB_SCALAR_FUNCTION_SET_ALIAS(ArraySortFun),
-	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayTransformFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayUniqueFun),
 	DUCKDB_SCALAR_FUNCTION(ArrayValueFun),
 	DUCKDB_SCALAR_FUNCTION(ASCIIFun),
@@ -179,7 +175,6 @@ static const StaticFunctionDefinition core_functions[] = {
 	DUCKDB_SCALAR_FUNCTION(ExpFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(FactorialFun),
 	DUCKDB_AGGREGATE_FUNCTION(FAvgFun),
-	DUCKDB_SCALAR_FUNCTION_ALIAS(FilterFun),
 	DUCKDB_SCALAR_FUNCTION(ListFlattenFun),
 	DUCKDB_SCALAR_FUNCTION_SET(FloorFun),
 	DUCKDB_SCALAR_FUNCTION(FormatFun),
@@ -234,34 +229,28 @@ static const StaticFunctionDefinition core_functions[] = {
 	DUCKDB_AGGREGATE_FUNCTION(ListFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ListAggrFun),
 	DUCKDB_SCALAR_FUNCTION(ListAggregateFun),
-	DUCKDB_SCALAR_FUNCTION_ALIAS(ListApplyFun),
 	DUCKDB_SCALAR_FUNCTION_SET(ListCosineDistanceFun),
 	DUCKDB_SCALAR_FUNCTION_SET(ListCosineSimilarityFun),
 	DUCKDB_SCALAR_FUNCTION_SET(ListDistanceFun),
 	DUCKDB_SCALAR_FUNCTION(ListDistinctFun),
 	DUCKDB_SCALAR_FUNCTION_SET_ALIAS(ListDotProductFun),
-	DUCKDB_SCALAR_FUNCTION(ListFilterFun),
 	DUCKDB_SCALAR_FUNCTION_SET(ListGradeUpFun),
 	DUCKDB_SCALAR_FUNCTION(ListHasAllFun),
 	DUCKDB_SCALAR_FUNCTION(ListHasAnyFun),
 	DUCKDB_SCALAR_FUNCTION_SET(ListInnerProductFun),
 	DUCKDB_SCALAR_FUNCTION_SET_ALIAS(ListNegativeDotProductFun),
 	DUCKDB_SCALAR_FUNCTION_SET(ListNegativeInnerProductFun),
-	DUCKDB_SCALAR_FUNCTION_SET_ALIAS(ListPackFun),
 	DUCKDB_SCALAR_FUNCTION_SET(ListReduceFun),
 	DUCKDB_SCALAR_FUNCTION_SET(ListReverseSortFun),
 	DUCKDB_SCALAR_FUNCTION_SET(ListSliceFun),
 	DUCKDB_SCALAR_FUNCTION_SET(ListSortFun),
-	DUCKDB_SCALAR_FUNCTION(ListTransformFun),
 	DUCKDB_SCALAR_FUNCTION(ListUniqueFun),
-	DUCKDB_SCALAR_FUNCTION_SET(ListValueFun),
 	DUCKDB_AGGREGATE_FUNCTION_SET_ALIAS(ListaggFun),
 	DUCKDB_SCALAR_FUNCTION(LnFun),
 	DUCKDB_SCALAR_FUNCTION_SET(LogFun),
 	DUCKDB_SCALAR_FUNCTION(Log10Fun),
 	DUCKDB_SCALAR_FUNCTION(Log2Fun),
 	DUCKDB_SCALAR_FUNCTION(LpadFun),
-	DUCKDB_SCALAR_FUNCTION_SET(LtrimFun),
 	DUCKDB_AGGREGATE_FUNCTION_SET(MadFun),
 	DUCKDB_SCALAR_FUNCTION_SET(MakeDateFun),
 	DUCKDB_SCALAR_FUNCTION(MakeTimeFun),
@@ -331,7 +320,6 @@ static const StaticFunctionDefinition core_functions[] = {
 	DUCKDB_SCALAR_FUNCTION(RightGraphemeFun),
 	DUCKDB_SCALAR_FUNCTION_SET(RoundFun),
 	DUCKDB_SCALAR_FUNCTION(RpadFun),
-	DUCKDB_SCALAR_FUNCTION_SET(RtrimFun),
 	DUCKDB_SCALAR_FUNCTION_SET(SecondsFun),
 	DUCKDB_AGGREGATE_FUNCTION(StandardErrorOfTheMeanFun),
 	DUCKDB_SCALAR_FUNCTION(SetBitFun),
@@ -385,7 +373,6 @@ static const StaticFunctionDefinition core_functions[] = {
 	DUCKDB_SCALAR_FUNCTION_SET(ToYearsFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(TransactionTimestampFun),
 	DUCKDB_SCALAR_FUNCTION(TranslateFun),
-	DUCKDB_SCALAR_FUNCTION_SET(TrimFun),
 	DUCKDB_SCALAR_FUNCTION_SET(TruncFun),
 	DUCKDB_SCALAR_FUNCTION(CurrentTransactionIdFun),
 	DUCKDB_SCALAR_FUNCTION(TypeOfFun),
@@ -395,7 +382,6 @@ static const StaticFunctionDefinition core_functions[] = {
 	DUCKDB_SCALAR_FUNCTION(UnionExtractFun),
 	DUCKDB_SCALAR_FUNCTION(UnionTagFun),
 	DUCKDB_SCALAR_FUNCTION(UnionValueFun),
-	DUCKDB_SCALAR_FUNCTION(UnpivotListFun),
 	DUCKDB_SCALAR_FUNCTION(UrlDecodeFun),
 	DUCKDB_SCALAR_FUNCTION(UrlEncodeFun),
 	DUCKDB_SCALAR_FUNCTION(UUIDFun),
