+++ extension/core_functions/scalar/date/functions.json
@@ -22,6 +22,14 @@
         "aliases": ["datediff"]
     },
     {
+        "name": "date_sub",
+        "parameters": "part,startdate,enddate",
+        "description": "The number of complete partitions between the timestamps",
+        "example": "date_sub('hour', TIMESTAMPTZ '1992-09-30 23:59:59', TIMESTAMPTZ '1992-10-01 01:58:00')",
+        "type": "scalar_function_set",
+        "aliases": ["datesub"]
+    },
+    {
         "name": "date_part",
         "parameters": "ts",
         "description": "Get subfield (equivalent to extract)",
@@ -30,12 +38,11 @@
         "aliases": ["datepart"]
     },
     {
-        "name": "date_sub",
-        "parameters": "part,startdate,enddate",
-        "description": "The number of complete partitions between the timestamps",
-        "example": "date_sub('hour', TIMESTAMPTZ '1992-09-30 23:59:59', TIMESTAMPTZ '1992-10-01 01:58:00')",
-        "type": "scalar_function_set",
-        "aliases": ["datesub"]
+        "name": "timezone",
+        "parameters": "ts",
+        "description": "Extract the timezone component from a date or timestamp",
+        "example": "timezone(timestamp '2021-08-03 11:59:44.123456')",
+        "type": "scalar_function_set"
     },
     {
         "name": "date_trunc",
@@ -304,13 +311,6 @@
         "parameters": "bucket_width,timestamp,origin",
         "description": "Truncate TIMESTAMPTZ by the specified interval bucket_width. Buckets are aligned relative to origin TIMESTAMPTZ. The origin defaults to 2000-01-03 00:00:00+00 for buckets that do not include a month or year interval, and to 2000-01-01 00:00:00+00 for month and year buckets",
         "example": "time_bucket(INTERVAL '2 weeks', TIMESTAMP '1992-04-20 15:26:00-07', TIMESTAMP '1992-04-01 00:00:00-07')",
-        "type": "scalar_function_set"
-    },
-    {
-        "name": "timezone",
-        "parameters": "ts",
-        "description": "Extract the timezone component from a date or timestamp",
-        "example": "timezone(timestamp '2021-08-03 11:59:44.123456')",
         "type": "scalar_function_set"
     },
     {
