+++ extension/core_functions/scalar/list/CMakeLists.txt
@@ -2,10 +2,7 @@ add_library_unity(
   duckdb_core_functions_list
   OBJECT
   flatten.cpp
-  list_transform.cpp
   range.cpp
-  list_value.cpp
-  list_filter.cpp
   list_has_any_or_all.cpp
   list_aggregates.cpp
   list_distance.cpp
