+++ src/function/function_list.cpp
@@ -74,11 +74,14 @@ static const StaticFunctionDefinition function[] = {
 	DUCKDB_SCALAR_FUNCTION_SET(InternalDecompressStringFun),
 	DUCKDB_SCALAR_FUNCTION_SET_ALIAS(AddFun),
 	DUCKDB_AGGREGATE_FUNCTION_SET(AnyValueFun),
+	DUCKDB_SCALAR_FUNCTION_ALIAS(ApplyFun),
 	DUCKDB_AGGREGATE_FUNCTION_SET_ALIAS(ArbitraryFun),
+	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayApplyFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayCatFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayConcatFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayContainsFun),
 	DUCKDB_SCALAR_FUNCTION_SET(ArrayExtractFun),
+	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayFilterFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayHasFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayIndexofFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayIntersectFun),
@@ -86,6 +89,7 @@ static const StaticFunctionDefinition function[] = {
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayPositionFun),
 	DUCKDB_SCALAR_FUNCTION_SET_ALIAS(ArrayResizeFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ArraySelectFun),
+	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayTransformFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayWhereFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ArrayZipFun),
 	DUCKDB_SCALAR_FUNCTION_SET(BitLengthFun),
@@ -107,6 +111,7 @@ static const StaticFunctionDefinition function[] = {
 	DUCKDB_SCALAR_FUNCTION_SET_ALIAS(DivideFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(EndsWithFun),
 	DUCKDB_SCALAR_FUNCTION(ErrorFun),
+	DUCKDB_SCALAR_FUNCTION_ALIAS(FilterFun),
 	DUCKDB_SCALAR_FUNCTION_SET(FinalizeFun),
 	DUCKDB_AGGREGATE_FUNCTION_SET(FirstFun),
 	DUCKDB_SCALAR_FUNCTION(GetVariableFun),
@@ -117,20 +122,26 @@ static const StaticFunctionDefinition function[] = {
 	DUCKDB_SCALAR_FUNCTION_SET(LengthFun),
 	DUCKDB_SCALAR_FUNCTION_SET(LengthGraphemeFun),
 	DUCKDB_SCALAR_FUNCTION(LikeEscapeFun),
+	DUCKDB_SCALAR_FUNCTION_ALIAS(ListApplyFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ListCatFun),
 	DUCKDB_SCALAR_FUNCTION(ListConcatFun),
 	DUCKDB_SCALAR_FUNCTION(ListContainsFun),
 	DUCKDB_SCALAR_FUNCTION_SET_ALIAS(ListElementFun),
 	DUCKDB_SCALAR_FUNCTION_SET(ListExtractFun),
+	DUCKDB_SCALAR_FUNCTION(ListFilterFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ListHasFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(ListIndexofFun),
 	DUCKDB_SCALAR_FUNCTION(ListIntersectFun),
+	DUCKDB_SCALAR_FUNCTION_SET_ALIAS(ListPackFun),
 	DUCKDB_SCALAR_FUNCTION(ListPositionFun),
 	DUCKDB_SCALAR_FUNCTION_SET(ListResizeFun),
 	DUCKDB_SCALAR_FUNCTION(ListSelectFun),
+	DUCKDB_SCALAR_FUNCTION(ListTransformFun),
+	DUCKDB_SCALAR_FUNCTION_SET(ListValueFun),
 	DUCKDB_SCALAR_FUNCTION(ListWhereFun),
 	DUCKDB_SCALAR_FUNCTION(ListZipFun),
 	DUCKDB_SCALAR_FUNCTION(LowerFun),
+	DUCKDB_SCALAR_FUNCTION_SET(LtrimFun),
 	DUCKDB_SCALAR_FUNCTION(MapContainsFun),
 	DUCKDB_AGGREGATE_FUNCTION_SET(MaxFun),
 	DUCKDB_SCALAR_FUNCTION_SET(MD5Fun),
@@ -154,6 +165,7 @@ static const StaticFunctionDefinition function[] = {
 	DUCKDB_SCALAR_FUNCTION_SET_ALIAS(RegexpSplitToArrayFun),
 	DUCKDB_SCALAR_FUNCTION(RemapStructFun),
 	DUCKDB_SCALAR_FUNCTION(RowFun),
+	DUCKDB_SCALAR_FUNCTION_SET(RtrimFun),
 	DUCKDB_SCALAR_FUNCTION_SET(SHA1Fun),
 	DUCKDB_SCALAR_FUNCTION_SET(SHA256Fun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(SplitFun),
@@ -187,8 +199,10 @@ static const StaticFunctionDefinition function[] = {
 	DUCKDB_SCALAR_FUNCTION_SET(SubstringGraphemeFun),
 	DUCKDB_SCALAR_FUNCTION_SET_ALIAS(SubtractFun),
 	DUCKDB_SCALAR_FUNCTION(SuffixFun),
+	DUCKDB_SCALAR_FUNCTION_SET(TrimFun),
 	DUCKDB_SCALAR_FUNCTION_SET(TryStrpTimeFun),
 	DUCKDB_SCALAR_FUNCTION_ALIAS(UcaseFun),
+	DUCKDB_SCALAR_FUNCTION(UnpivotListFun),
 	DUCKDB_SCALAR_FUNCTION(UpperFun),
 	DUCKDB_SCALAR_FUNCTION_SET(VariantExtractFun),
 	DUCKDB_SCALAR_FUNCTION(VariantNormalizeFun),
