+++ src/function/scalar/list/CMakeLists.txt
@@ -6,7 +6,11 @@ add_library_unity(
   list_intersect.cpp
   list_resize.cpp
   list_zip.cpp
-  list_select.cpp)
+  list_select.cpp
+  list_filter.cpp
+  list_transform.cpp
+  list_value.cpp
+  lambda_functions.cpp)
 set(ALL_OBJECT_FILES
     ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:duckdb_func_list_nested>
     PARENT_SCOPE)
