+++ src/function/scalar/list/functions.json
@@ -9,6 +9,41 @@
         "aliases": ["array_select"]
     },
     {
+        "name": "list_transform",
+        "parameters": "list,lambda(x)",
+        "description": "Returns a list that is the result of applying the `lambda` function to each element of the input `list`. The return type is defined by the return type of the `lambda` function.",
+        "example": "list_transform([1, 2, 3], lambda x : x + 1)",
+        "categories": ["list", "lambda"],
+        "type": "scalar_function",
+        "aliases": ["array_transform", "list_apply", "array_apply", "apply"]
+    },
+    {
+        "name": "unpivot_list",
+        "parameters": "any,...",
+        "description": "Identical to list_value, but generated as part of unpivot for better error messages.",
+        "example": "unpivot_list(4, 5, 6)",
+        "categories": ["list"],
+        "type": "scalar_function"
+    },
+    {
+        "name": "list_filter",
+        "parameters": "list,lambda(x)",
+        "description": "Constructs a list from those elements of the input `list` for which the `lambda` function returns `true`. DuckDB must be able to cast the `lambda` function's return type to `BOOL`. The return type of `list_filter` is the same as the input list's.",
+        "example": "list_filter([3, 4, 5], lambda x : x > 4)",
+        "categories": ["list", "lambda"],
+        "type": "scalar_function",
+        "aliases": ["array_filter", "filter"]
+    },
+    {
+        "name": "list_value",
+        "parameters": "any,...",
+        "description": "Creates a LIST containing the argument values.",
+        "example": "list_value(4, 5, 6)",
+        "categories": ["list"],
+        "type": "scalar_function_set",
+        "aliases": ["list_pack"]
+    },
+    {
         "name": "list_where",
         "parameters": "value_list,mask_list",
         "description": "Returns a list with the `BOOLEAN`s in `mask_list` applied as a mask to the `value_list`.",
