+++ src/function/scalar/list/lambda_functions.cpp
