+++ src/function/scalar/list/list_transform.cpp
@@ -1,4 +1,4 @@
-#include "core_functions/scalar/list_functions.hpp"
+#include "duckdb/function/scalar/list_functions.hpp"
 
 #include "duckdb/function/lambda_functions.hpp"
 #include "duckdb/planner/expression/bound_cast_expression.hpp"
