+++ src/function/scalar/string/CMakeLists.txt
@@ -18,7 +18,8 @@ add_library_unity(
   strip_accents.cpp
   string_split.cpp
   suffix.cpp
-  contains.cpp)
+  contains.cpp
+  trim.cpp)
 set(ALL_OBJECT_FILES
     ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:duckdb_func_string_main>
     PARENT_SCOPE)
